+++ connect/api/v2/organizations/views.py
@@ -1,22 +1,31 @@
-from rest_framework.viewsets import GenericViewSet
 from rest_framework import mixins, status
+from rest_framework.viewsets import GenericViewSet
 from rest_framework.response import Response
-
-from connect.common.models import Organization, OrganizationAuthorization, OrganizationRole
-from connect.api.v2.organizations.serializers import OrganizationSeralizer
-from connect.api.v2.projects.serializers import ProjectSerializer
-
-from drf_yasg2.utils import swagger_auto_schema
-from connect.api.v2.organizations.api_schemas import (
-    create_organization_schema,
-)
-
 from rest_framework.permissions import IsAuthenticated
+from drf_yasg2.utils import swagger_auto_schema
+
+from connect.common.models import (
+    Organization,
+    OrganizationAuthorization,
+    OrganizationRole
+)
 from connect.api.v1.organization.permissions import (
     Has2FA,
     OrganizationHasPermission,
 )
+
 from connect.api.v2.permissions import OrgIPPermission
+
+from connect.api.v2.organizations.serializers import (
+    OrganizationSeralizer,
+    NestedAuthorizationOrganizationSerializer
+)
+from connect.api.v2.projects.serializers import ProjectSerializer
+from connect.api.v2.organizations.api_schemas import (
+    create_organization_schema,
+)
+from connect.api.v2.paginations import CustomCursorPagination
+
 
 
 class OrganizationViewSet(
@@ -30,7 +39,10 @@ class OrganizationViewSet(
     queryset = Organization.objects.all()
     serializer_class = OrganizationSeralizer
     lookup_field = "uuid"
+
     permission_classes = [IsAuthenticated, OrgIPPermission, OrganizationHasPermission, Has2FA]
+    pagination_class = CustomCursorPagination
+
 
     def get_queryset(self, *args, **kwargs):
         if getattr(self, "swagger_fake_view", False):
@@ -86,3 +98,15 @@ class OrganizationViewSet(
         user_email = self.request.user.email
         instance.perform_destroy_ai_organization(user_email)
         instance.delete()
+
+
+class OrganizationAuthorizationViewSet(
+    mixins.RetrieveModelMixin,
+    GenericViewSet,
+):
+    queryset = Organization.objects
+    permission_classes = [IsAuthenticated, OrganizationHasPermission]
+    lookup_field = "uuid"
+
+    def get_serializer_class(self):
+        return NestedAuthorizationOrganizationSerializer